Fix failing tests for small/large machine costs

+++ src/test/java/com/davidbarsky/dag/models/states/MachineTypeTest.java
@@ -8,12 +8,12 @@ class MachineTypeTest {
     @Test
     void getCostForSmallMachine() {
         MachineType smallMachine = MachineType.SMALL;
-        assertEquals(new Integer(2), smallMachine.getCost());
+        assertEquals(new Integer(1), smallMachine.getCost());
     }
 
     @Test
     void getCostForLargeMachine() {
         MachineType largeMachine = MachineType.LARGE;
-        assertEquals(new Integer(4), largeMachine.getCost());
+        assertEquals(new Integer(2), largeMachine.getCost());
     }
 }
